refactor accessor function to *accessor files
