+++ corehq/apps/reports/tasks.py
@@ -1,5 +1,4 @@
 from datetime import datetime, timedelta
-from dateutil.parser import parse
 import hashlib
 import os
 import tempfile
@@ -23,7 +22,6 @@ from couchexport.files import Temp
 from couchexport.groupexports import export_for_group, rebuild_export
 from couchexport.tasks import cache_file_to_be_served
 from couchforms.analytics import app_has_been_submitted_to_in_last_30_days
-from couchforms.models import XFormInstance
 from dimagi.utils.couch.cache.cache_core import get_redis_client
 from dimagi.utils.django.email import send_HTML_email
 from dimagi.utils.logging import notify_exception
@@ -46,6 +44,12 @@ from corehq.pillows.mappings.app_mapping import APP_INDEX
 from corehq.util.files import file_extention_from_filename
 from corehq.util.view_utils import absolute_reverse
 
+from .analytics.couchaccessors import (
+    get_form_ids_having_multimedia as couch_get_form_ids_having_multimedia
+)
+from .analytics.esaccessors import (
+    get_form_ids_having_multimedia as es_get_form_ids_having_multimedia
+)
 from .dbaccessors import get_all_hq_group_export_configs
 from .export import save_metadata_export_to_tempfile
 from .models import (
@@ -381,13 +385,6 @@ def _get_export_properties(export_id, export_is_legacy):
                         properties.add("-".join(path_parts))
     return properties
 
-def _get_attachment_dicts_from_form(form):
-    if 'external_blobs' in form:
-        return form['external_blobs'].values()
-    elif '_attachments' in form:
-        return form['_attachments'].values()
-    return []
-
 
 def _get_form_ids_having_multimedia(domain, app_id, xmlns, startdate, enddate, export_is_legacy):
     """
@@ -395,35 +392,17 @@ def _get_form_ids_having_multimedia(domain, app_id, xmlns, startdate, enddate, e
     Each form has a multimedia attachment and meets the given filters.
     """
     if not export_is_legacy:
-        # TODO: Remove references to _attachments once all forms have been migrated to Riak
-        query = (FormES()
-                 .domain(domain)
-                 .app(app_id)
-                 .xmlns(xmlns)
-                 .submitted(gte=parse(startdate), lte=parse(enddate))
-                 .remove_default_filter("has_user")
-                 .source(['_attachments', '_id', 'external_blobs']))
-        form_ids = set()
-        for form in query.scroll():
-            try:
-                for attachment in _get_attachment_dicts_from_form(form):
-                    if attachment['content_type'] != "text/xml":
-                        form_ids.add(form['_id'])
-                        continue
-            except AttributeError:
-                pass
+        fetch_fn = es_get_form_ids_having_multimedia
     else:
-        key = [domain, app_id, xmlns]
-        form_ids = {
-            f['id'] for f in
-            XFormInstance.get_db().view(
-                "attachments/attachments",
-                start_key=key + [startdate],
-                end_key=key + [enddate, {}],
-                reduce=False
-            )
-        }
-    return form_ids
+        fetch_fn = couch_get_form_ids_having_multimedia
+
+    return fetch_fn(
+        domain,
+        app_id,
+        xmlns,
+        startdate,
+        enddate,
+    )
 
 
 def _extract_form_attachment_info(form, properties):
